requirements.txt
pillow~=10.4.0
openpyxl~=3.1.5
docxtpl~=0.18.0
flask~=3.0.3